requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
python-dotenv>=0.19.0
pexels-api-python==0.1.4
Pillow>=10.0.0
requests>=2.31.0
python-multipart>=0.0.6
gTTS>=2.3.1
supabase>=2.0.0
